% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{district_age_median}
\alias{district_age_median}
\title{Median age}
\format{
A data frame with 1539 rows and 4 variables:
\describe{
\item{\code{sex}}{character. Sex}
\item{\code{year}}{integer. Year, 2011-2019}
\item{\code{median_age}}{double. Median age.}
\item{\code{KOD_ZUJ}}{character. district code (ZUJ code, CZSO codelist 51 http://apl.czso.cz/iSMS/cisdet.jsp?kodcis=51)}
}
}
\usage{
district_age_median
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
