% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{district_age_structure}
\alias{district_age_structure}
\title{District population by age and sex}
\format{
A data frame with 9747 rows and 5 variables:
\describe{
\item{\code{KOD_ZUJ}}{character. district code (ZUJ code, CZSO codelist 51 http://apl.czso.cz/iSMS/cisdet.jsp?kodcis=51)}
\item{\code{KOD_SO}}{character. admin. district code (CZSO codelist 72 - CISSOP http://apl.czso.cz/iSMS/cisdet.jsp?kodcis=72)}
\item{\code{count}}{double. number of residents.}
\item{\code{age}}{integer. 5-year age group, except 85+.}
\item{\code{year}}{integer. Year; data for 31. 12.}
\item{\code{sex}}{character. Sex.}
}
}
\usage{
district_age_structure
}
\description{
Population of each of the 57 districts by 5-year age groups and sex, 2011-2019, from the Czech Statistical Office
}
\details{
Source: https://www.czso.cz/csu/xa/casove-rady-za-mestske-casti-prahy
}
\keyword{datasets}
