% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{district_geofacet}
\alias{district_geofacet}
\title{Dataset to be used in the {geofacet} package}
\format{
A data frame with 57 rows and 11 variables:
\describe{
\item{\code{code}}{character. RUIAN code. Normally should serve as ID to distribute your data points into grid cells.}
\item{\code{label}}{character. Three-character unique label.}
\item{\code{row}}{integer. row number.}
\item{\code{col}}{integer. column number.}
}
}
\usage{
district_geofacet
}
\description{
Use this as the \code{grid} argument to \code{geofacet::facet_geo()}. The layout corresponds to \code{district_tilegram}.
}
\seealso{
Other Mapping: 
\code{\link{district_hexogram}},
\code{\link{district_names}},
\code{\link{district_tilegram}},
\code{\link{prg_basemap}()},
\code{\link{prg_endpoints}},
\code{\link{prg_tile}()}
}
\concept{Mapping}
\keyword{datasets}
