% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{district_hexogram}
\alias{district_hexogram}
\title{Equal-area hexogram of Prague districts}
\format{A data frame with 57 rows and 20 variables:}
\usage{
district_hexogram
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
