% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{district_names}
\alias{district_names}
\title{Table of several variants of each districts names, with code}
\format{
A data frame with 57 rows and 11 variables:
\describe{
\item{\code{kod}}{character. RUIAN code.}
\item{\code{label}}{character. Three-character unique label.}
\item{\code{nazev}}{integer. full name.}
\item{\code{name_medium}}{integer. shortened name.}
\item{\code{name_short}}{integer. short name.}
\item{\code{name_spaced}}{integer. full name with spaced dash.}
}
}
\usage{
district_names
}
\description{
For when a shorter label is needed or for linking with other data by name.
}
\seealso{
Other Mapping: 
\code{\link{district_geofacet}},
\code{\link{district_hexogram}},
\code{\link{district_tilegram}},
\code{\link{prg_basemap}()},
\code{\link{prg_endpoints}},
\code{\link{prg_tile}()}
}
\concept{Mapping}
\keyword{datasets}
