% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{district_tilegram}
\alias{district_tilegram}
\title{Equal-area tilegram of Prague districts}
\format{
A data frame with 57 rows and 11 variables:
\describe{
\item{\code{kod}}{character. RUIAN code.}
\item{\code{nazev}}{character. Full name.}
\item{\code{label}}{character. Three-character unique label.}
\item{\code{mop_kod}}{character. Prague district code.}
\item{\code{spo_kod}}{character. Admin district code.}
\item{\code{pou_kod}}{character. Higher-level unit code.}
\item{\code{okres_kod}}{character. District code.}
\item{\code{CENTROIX}}{double. X coordinate of centroid.}
\item{\code{CENTROIY}}{double. Y coordinate of centroid.}
\item{\code{row}}{integer. row number.}
\item{\code{col}}{integer. column number.}
\item{\code{geometry}}{list. geometry.}
}
}
\usage{
district_tilegram
}
\description{
Equal-area tilegram of 57 Prague districts. Best available representation
though some district tiles neighbor districts they do not in reality.
}
\seealso{
Other Mapping: 
\code{\link{district_hexogram}},
\code{\link{prg_basemap}()},
\code{\link{prg_endpoints}},
\code{\link{prg_tile}()}
}
\concept{Mapping}
\keyword{datasets}
