% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cuzk_prague.R
\name{get_cuzk_prague}
\alias{get_cuzk_prague}
\title{Download GML geodata file for Prague from CUZK}
\usage{
get_cuzk_prague(
  date = strftime(lubridate::rollback(Sys.Date()), format = "\%Y\%m\%d"),
  path = ".",
  cachetime = 32
)
}
\arguments{
\item{date}{Date of file, must be last day of month. Defaults to end of previous month.}

\item{path}{Where to save the downloaded zip file and unzipped xml file}

\item{cachetime}{Redownload if the zip file found in path exists and is older than cachetime. Defaults to 32. Set to 0 to always redownload. Only checks existence and timestamp of downloaded zip file, not unzipped file.}
}
\value{
string Path to downloaded and unzipped file incl. path
}
\description{
Download GML geodata file for Prague from CUZK
}
\keyword{internal}
