% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pragr-package.R
\docType{package}
\name{pragr-package}
\alias{pragr}
\alias{pragr-package}
\title{pragr: Tools for visualising Prague data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A bridge between Prague geodata and R to enable
    visualisation. Currently, it provides access to raster map layers
    provided by the Prague geoportal and several utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petrbouchal/pragr}
  \item Report bugs at \url{https://github.com/petrbouchal/pragr/issues}
}

}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com} (\href{https://orcid.org/0000-0002-0471-716X}{ORCID})

}
\keyword{internal}
