% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap_prg.R
\name{prg_basemap}
\alias{prg_basemap}
\title{Prague base maps for ggplot2}
\usage{
prg_basemap(data, image_service = "orto", layer = "", width = 900,
  alpha = 1, buffer = 0, verbose = F)
}
\arguments{
\item{data}{sf data frame from which to extract the bounding box}

\item{image_service}{map service from which to draw the map; \code{prg_endpoints} provides details.}

\item{layer}{layer from map service to use, see https://mpp.praha.eu/arcgis/rest/services/}

\item{width}{width in pixels, in effect sets image resolution; integer or "max"}

\item{alpha}{transparency of the tile}

\item{buffer}{distance between feature end and image end; for EPSG 5514 in meters.}

\item{verbose}{display information on image URLs and image processing.}
}
\value{
list including raster annotation layers for ggplot2
}
\description{
Include raster maps from Prague geoportal in your ggplot map
}
\examples{
\dontrun{
praha1 <- CzechData::load_RUIAN_settlement(prg_kod, "MOMC_P", WGS84 = F) \%>\%
  filter(nazev == 'Praha 1')

ggplot() +
  prg_basemap(data = praha1, alpha = .8, buffer = 200,
              image_service = 'mapy_archiv', layer = 6) +
  geom_sf(data = praha1, fill = alpha("red", 0.6), colour = NA) +
  theme_void()
}
}
\seealso{
Other Mapping: \code{\link{prg_endpoints}},
  \code{\link{prg_tile}}
}
\concept{Mapping}
