% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_bbox_krovak}
\alias{prg_bbox_krovak}
\title{Prague bbox in EPSG 5514 (Krovak)}
\format{
named numeric of class \code{bbox}, length 4
\describe{
Prague RUIAN code
}
}
\usage{
prg_bbox_krovak
}
\description{
Prague bbox in EPSG 5514 (Krovak)
}
\seealso{
Other Codes and metadata: 
\code{\link{prg_bbox_wgs84}},
\code{\link{prg_fua_oecd}},
\code{\link{prg_ico}},
\code{\link{prg_kod}},
\code{\link{prg_kraj}},
\code{\link{prg_lau1}},
\code{\link{prg_metro_oecd}},
\code{\link{prg_nuts2}},
\code{\link{prg_nuts3}},
\code{\link{prg_okres_nuts}},
\code{\link{prg_okres}}
}
\concept{Codes and metadata}
\keyword{datasets}
