% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_bbox_wgs84}
\alias{prg_bbox_wgs84}
\title{Prague bbox in EPSG 4326 (WGS 84)}
\format{named numeric of class \code{bbox}, length 4
\describe{
Prague RUIAN code
}}
\usage{
prg_bbox_wgs84
}
\description{
Prague bbox in EPSG 4326 (WGS 84)
}
\seealso{
Other Codes and metadata: 
\code{\link{prg_bbox_krovak}},
\code{\link{prg_fua_oecd}},
\code{\link{prg_ico}},
\code{\link{prg_kod}},
\code{\link{prg_lau1}},
\code{\link{prg_metro_oecd}},
\code{\link{prg_nuts2}},
\code{\link{prg_nuts3}},
\code{\link{prg_okres_nuts}},
\code{\link{prg_okres}}
}
\concept{Codes and metadata}
\keyword{datasets}
