% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prg_endpoints.R
\docType{data}
\name{prg_endpoints}
\alias{prg_endpoints}
\title{IPR Map Services}
\format{
A data frame with 18 rows and 5 variables:
\describe{
\item{\code{name}}{character. Name of the service, to be used as parameter in prg_tile() and prg_basemap()}
\item{\code{type}}{character. "tile" or "image"}
\item{\code{description}}{character. Description of the map source}
\item{\code{endpoint}}{character. endpoint. Used internally}
\item{\code{url}}{character. URL, used internally but also points to human readable interface.}
}
}
\usage{
prg_endpoints
}
\description{
Contains information on a selection of available Prague raster map services.
Can be used manually to inspect services and is used internally by
prg_tile() and prg_basemap() for retrieving data from servers.
}
\seealso{
Other Mapping: 
\code{\link{district_hexogram}},
\code{\link{prg_basemap}()},
\code{\link{prg_tile}()}
}
\concept{Mapping}
\keyword{datasets}
