% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_fua_oecd}
\alias{prg_fua_oecd}
\title{Prague code in OECD database of Functional Urban Areas}
\format{character vector of length 1
\describe{
Prague code in OECD Functional Urban Area database.
}}
\usage{
prg_fua_oecd
}
\description{
Prague code in OECD Functional Urban Area database.
}
\details{
See https://www.oecd.org/cfe/regional-policy/functionalurbanareasbycountry.htmz
}
\keyword{datasets}
