% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_lau1}
\alias{prg_lau1}
\title{Prague LAU1 code}
\format{character vector of length 1
\describe{
Prague LAU1 code. Comes from CZSO registry 109 - OKRES_LAU.
}}
\usage{
prg_lau1
}
\description{
Prague LAU1 code - level of okres.
}
\keyword{datasets}
