% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_metro_oecd}
\alias{prg_metro_oecd}
\title{Prague code in OECD CITIES database of data on Functional Urban Areas}
\format{character vector of length 1
\describe{
Prague code in OECD Functional Urban Area database.
}}
\usage{
prg_metro_oecd
}
\description{
Prague code in OECD database on Metropolitan Areas.
}
\details{
See https://stats.oecd.org/Index.aspx?DataSetCode=CITIES
}
\keyword{datasets}
