% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_nuts2}
\alias{prg_nuts2}
\title{Prague NUTS2 code}
\format{
character vector of length 1
\describe{
Prague NUTS2 code
}
}
\usage{
prg_nuts2
}
\description{
Prague NUTS2 code
}
\seealso{
Other Codes and metadata: 
\code{\link{prg_bbox_krovak}},
\code{\link{prg_bbox_wgs84}},
\code{\link{prg_fua_oecd}},
\code{\link{prg_ico}},
\code{\link{prg_kod}},
\code{\link{prg_lau1}},
\code{\link{prg_metro_oecd}},
\code{\link{prg_nuts3}},
\code{\link{prg_okres_nuts}},
\code{\link{prg_okres}}
}
\concept{Codes and metadata}
\keyword{datasets}
