% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_nuts3}
\alias{prg_nuts3}
\title{Prague NUTS3 code}
\format{character vector of length 1
\describe{
Prague NUTS3 code
}}
\usage{
prg_nuts3
}
\description{
Prague NUTS3 code, level of kraj. Comes from CZSO registry 108 - NUTS3-2004.
}
\keyword{datasets}
