% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_okres}
\alias{prg_okres}
\title{Prague 'okres' code}
\format{
character vector of length 1
\describe{
Prague okres code
}
}
\usage{
prg_okres
}
\description{
Prague 'okres' code. Comes from CZSO registry 101 - OKRES_NUTS.
}
\seealso{
Other Codes and metadata: 
\code{\link{prg_bbox_krovak}},
\code{\link{prg_bbox_wgs84}},
\code{\link{prg_fua_oecd}},
\code{\link{prg_ico}},
\code{\link{prg_kod}},
\code{\link{prg_kraj}},
\code{\link{prg_lau1}},
\code{\link{prg_metro_oecd}},
\code{\link{prg_nuts2}},
\code{\link{prg_nuts3}},
\code{\link{prg_okres_nuts}}
}
\concept{Codes and metadata}
\keyword{datasets}
