% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_okres_nuts}
\alias{prg_okres_nuts}
\title{Prague 'okres LAU' code}
\format{character vector of length 1
\describe{
Prague okres (NUTS) code.
}}
\usage{
prg_okres_nuts
}
\description{
Prague 'okres LAU' code. Comes from CZSO register 101 (OKRES_NUTS) - yes these are
confusing but that is how it is.
}
\seealso{
Other Codes and metadata: \code{\link{prg_bbox_krovak}},
  \code{\link{prg_bbox_wgs84}}, \code{\link{prg_fua_oecd}},
  \code{\link{prg_ico}}, \code{\link{prg_kod}},
  \code{\link{prg_lau1}}, \code{\link{prg_metro_oecd}},
  \code{\link{prg_nuts2}}, \code{\link{prg_nuts3}},
  \code{\link{prg_okres}}
}
\concept{Codes and metadata}
\keyword{datasets}
