% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_prg.R
\name{prg_tile}
\alias{prg_tile}
\title{Prague tiles for ggplot2 plots}
\usage{
prg_tile(data, tile_service, zoom = 6, alpha = 1, buffer = 0)
}
\arguments{
\item{data}{DESCRIPTION.}

\item{tile_service}{DESCRIPTION.}

\item{zoom}{DESCRIPTION.}

\item{alpha}{DESCRIPTION.}

\item{buffer}{DESCRIPTION.}
}
\value{
annotation object for ggplot
}
\description{
Include raster tiles for Prague in ggplot2
}
