% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pragr-package.R
\docType{package}
\name{pragr-package}
\alias{pragr}
\alias{pragr-package}
\title{pragr: Tools for visualising Prague data}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

This package aims to make Prague geodata available in R for
    visualisation. Currently, it provides access to raster map layers
    provided by the Prague geoportal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petrbouchal/pragr}
  \item Report bugs at \url{https://github.com/petrbouchal/pragr/issues}
}

}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com}

}
\keyword{internal}
