% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap_prg.R
\name{prg_basemap}
\alias{prg_basemap}
\title{Prague base maps for ggplot2}
\usage{
prg_basemap(data, image_service = "ortofoto", layer = "", size = 900,
  alpha = 1, buffer = 0)
}
\arguments{
\item{data}{sf data frame from which to extract the bounding box}

\item{image_service}{map service from which to draw the map; \code{show_map_services()} provides a list.}

\item{layer}{layer from map service to use, see https://mpp.praha.eu/arcgis/rest/services/}

\item{size}{longer side of the downloaded image in pixels}

\item{alpha}{transparency}

\item{buffer}{distance between feature end and image end; for EPSG 5514 in meters.}
}
\value{
annotation object for ggplot
}
\description{
Include raster maps from Prague geoportal in your ggplot map
}
