% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_kraj}
\alias{prg_kraj}
\title{Prague 'kraj' code}
\format{
character vector of length 1
\describe{
Prague kraj code
}
#' @family Codes and metadata
}
\usage{
prg_kraj
}
\description{
Prague 'kraj' code. Comes from CZSO registry 100, KRAJ_NUTS.
}
\keyword{datasets}
