% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_lau1}
\alias{prg_lau1}
\title{Prague LAU1 code}
\format{
character vector of length 1
\describe{
Prague LAU1 code. Comes from CZSO registry 109 - OKRES_LAU.
}
}
\usage{
prg_lau1
}
\description{
Prague LAU1 code - level of okres.
}
\seealso{
Other Codes and metadata: 
\code{\link{prg_bbox_krovak}},
\code{\link{prg_bbox_wgs84}},
\code{\link{prg_fua_oecd}},
\code{\link{prg_ico}},
\code{\link{prg_kod}},
\code{\link{prg_kraj}},
\code{\link{prg_metro_oecd}},
\code{\link{prg_nuts2}},
\code{\link{prg_nuts3}},
\code{\link{prg_okres_nuts}},
\code{\link{prg_okres}}
}
\concept{Codes and metadata}
\keyword{datasets}
