% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_metro_oecd}
\alias{prg_metro_oecd}
\title{Prague code in OECD CITIES database of data on Functional Urban Areas}
\format{
character vector of length 1
\describe{
Prague code in OECD Functional Urban Area database.
}
}
\usage{
prg_metro_oecd
}
\description{
Prague code in OECD database on Metropolitan Areas.
}
\details{
See https://stats.oecd.org/Index.aspx?DataSetCode=CITIES
}
\seealso{
Other Codes and metadata: 
\code{\link{prg_bbox_krovak}},
\code{\link{prg_bbox_wgs84}},
\code{\link{prg_fua_oecd}},
\code{\link{prg_ico}},
\code{\link{prg_kod}},
\code{\link{prg_lau1}},
\code{\link{prg_nuts2}},
\code{\link{prg_nuts3}},
\code{\link{prg_okres_nuts}},
\code{\link{prg_okres}}
}
\concept{Codes and metadata}
\keyword{datasets}
