% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_nuts2}
\alias{prg_nuts2}
\title{Prague NUTS2 code}
\format{character vector of length 1
\describe{
Prague NUTS2 code
}}
\usage{
prg_nuts2
}
\description{
Prague NUTS2 code
}
\keyword{datasets}
