% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{prg_okres}
\alias{prg_okres}
\title{Prague 'okres' code}
\format{character vector of length 1
\describe{
Prague okres code
}}
\usage{
prg_okres
}
\description{
Prague 'okres' code. Comes from CZSO registry 101 - OKRES_NUTS.
}
\keyword{datasets}
