% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_prg.R
\name{prg_tile}
\alias{prg_tile}
\title{Prague tiles for ggplot2 plots}
\usage{
prg_tile(data, tile_service, zoom = 6, alpha = 1, buffer = 0,
  verbose = F)
}
\arguments{
\item{data}{sf data frame from which to extract the bounding box}

\item{tile_service}{map service from which to draw the map (see \code{prg_endpoints}), or a URL of a service.}

\item{zoom}{zoom level, from 0 to the service's limit}

\item{alpha}{transparency of the tiles.}

\item{buffer}{distance between feature end and tile end; for EPSG 5514 in meters.}

\item{verbose}{display information on tile URLs and image processing.}
}
\value{
list including raster annotation layers for ggplot2
}
\description{
Include raster tiles for Prague in ggplot2
}
