% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_prg.R
\name{prg_tile}
\alias{prg_tile}
\title{Prague tiles for ggplot2 plots}
\usage{
prg_tile(
  data,
  tile_service = "orto",
  zoom = 6,
  alpha = 1,
  buffer = 0,
  verbose = F
)
}
\arguments{
\item{data}{sf data frame from which to extract the bounding box}

\item{tile_service}{map service from which to draw the map (see \code{prg_endpoints}), or a URL of a service.}

\item{zoom}{zoom level, from 0 to the service's limit}

\item{alpha}{transparency of the tiles.}

\item{buffer}{distance between feature end and tile end; for EPSG 5514 in meters.}

\item{verbose}{display information on tile URLs and image processing.}
}
\value{
list including raster annotation layers for ggplot2
}
\description{
Include raster tiles for Prague in ggplot2
}
\examples{
\dontrun{
praha1 <- CzechData::load_RUIAN_settlement(prg_kod, "MOMC_P", WGS84 = F) \%>\%
  filter(nazev == 'Praha 1')

ggplot() +
prg_tile(data = praha1, zoom = 10, alpha = .7, buffer = 200,
         tile_service = 'orto') +
  geom_sf(data = praha1, fill = alpha("red", 0.6), colour = NA) +
  theme_void()
}

}
\seealso{
Other Mapping: 
\code{\link{district_geofacet}},
\code{\link{district_hexogram}},
\code{\link{district_names}},
\code{\link{district_tilegram}},
\code{\link{prg_basemap}()},
\code{\link{prg_endpoints}}
}
\concept{Mapping}
