% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap_prg.R
\name{show_map_services}
\alias{show_map_services}
\title{Show map services available in \code{basemap_prg}}
\usage{
show_map_services()
}
\value{
a data frame with name and URL
}
\description{
Show map services available in \code{basemap_prg}
}
